twitter service

+++ core_java/twitter/src/main/java/ca/jrvs/apps/twitter/model/Tweet.java
@@ -105,7 +105,7 @@ public class Tweet {
   }
 
   @JsonProperty("retweet_count")
-  public void setRetweetCount(int retweetCount) {
+  public void setRetweetCount(long retweetCount) {
     this.retweetCount = retweetCount;
   }
 
@@ -115,7 +115,7 @@ public class Tweet {
   }
 
   @JsonProperty("favorite_count")
-  public void setFavoriteCount(int favoriteCount) {
+  public void setFavoriteCount(long favoriteCount) {
     this.favoriteCount = favoriteCount;
   }
